ENH: new Qt GUI testing framework
This allows developers to record a test by interacting with the GUI as a user
would.  The test is saved as .xml and can be played back either manually or
via ctest.  Currently this is only built into TraceEdit, but the bulk of the
framework is located in ftkGUI.  Therefore, any application using this library
should be able to include this testing framework with minimal difficulty.



git-svn-id: https://farsight-svn.ee.uh.edu/repos/farsight/trunk@3375 9053f808-4e3f-0410-9672-ef3bfcbbeb84