TraceEdit: Use vtkSmartVolumeMapper.
TraceEdit was optionally using one of two volume mappers to do its volume
rendering: vtkOpenGLVolumeTextureMapper3D or vtkOpenGLGPUVolumeRayCastMapper.

The RayCastMapper is much faster, and was only available by setting USE_GPUREN
to ON in the CMake configuration.  However, not all hardware supportws the
RayCastMapper.  Not all hardware supports the OpenGLVolumeTextureMapper3D
either, which resulted in a render window without any valid content.

In this patch we replace both with a vtkSmartVolumeMapper, which will choose the
fastest volume mapper supported for the present hardware for both interactive
and non-interactive rendering.

git-svn-id: https://farsight-svn.ee.uh.edu/repos/farsight/trunk@3993 9053f808-4e3f-0410-9672-ef3bfcbbeb84