Add a ProcessObjectProgressUpdater class.
This class is used to update a QProgressBar and QLabel display the progress of
an itk::ProcessObject, e.g. an itk::ImageToImageFilter.

It is applied to the TraceEdit's itk::ImageFileReader's.

The progress bar will only be effective if the ITK filter in question emits
itk::ProgressEvents's.  Unfortuately, that does not appear to be the case for
itk::ImageFileReader.

git-svn-id: https://farsight-svn.ee.uh.edu/repos/farsight/trunk@3995 9053f808-4e3f-0410-9672-ef3bfcbbeb84