Updates for building against VTK 6.
The USE_VTK6 preprocessor macro was remove with preference to checking the
version of VTK_VERSION_MAJOR to reduce preprocessor macro bloat.  A number of
other fixes were made for VTK6.

The ItkVtkGlue was updated to is in current ITK master.  In the future, this
module should be removed entirely from the repository and the one in ITK used
instead to prevent an unnecessary maintenance burden.

git-svn-id: https://farsight-svn.ee.uh.edu/repos/farsight/trunk@4018 9053f808-4e3f-0410-9672-ef3bfcbbeb84